pythonchallenge/requirements.txt
requests
lxml
pillow